Refactor(modules): The project was organized
